changes new

+++ src/main/java/com/stocks/services/StockService.java
@@ -1,6 +1,7 @@
 package com.stocks.services;
 
 import com.stocks.entities.StockEntity;
+import com.stocks.exceptions.RecordNotFoundException;
 import com.stocks.models.Stock;
 import com.stocks.repositories.StocksRepository;
 import org.dozer.DozerBeanMapper;
@@ -13,6 +14,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -21,18 +23,16 @@ public class StockService {
     @Autowired
     private StocksRepository stocksRepository;
 
-    public Stock addStock(Stock stock) {
+    public Stock addStock(Stock stock) throws  Exception{
         StockEntity stockEntity = new DozerBeanMapper().map(stock, StockEntity.class);
         stockEntity = stocksRepository.save(stockEntity);
         Stock stockSaved = new DozerBeanMapper().map(stockEntity, Stock.class);
         return stockSaved;
     }
 
-    public List<Stock> getStocks(Integer pageNo, Integer pageSize, String sortBy) {
+    public List<Stock> getStocks(Integer pageNo, Integer pageSize, String sortBy) throws  Exception{
         Pageable paging = PageRequest.of(pageNo, pageSize, Sort.by(sortBy));
-
         Page<StockEntity> stockEntityList = (Page<StockEntity>) stocksRepository.findAll(paging);
-
         if(stockEntityList.hasContent()) {
             ArrayList stockList = new DozerBeanMapper().map(stockEntityList.getContent(), new ArrayList<Stock>().getClass());
             return stockList;
@@ -42,21 +42,29 @@ public class StockService {
 
     }
 
-    public Stock getStockByStockId(Integer stockId) {
+    public Stock getStockByStockId(Integer stockId) throws  Exception{
         Optional<StockEntity> stockEntityOption = stocksRepository.findById(stockId);
+        if(Objects.nonNull(stockEntityOption))
+            throw new RecordNotFoundException("Record not found");
         Stock stock = new DozerBeanMapper().map(stockEntityOption.get(), Stock.class);
         return stock;
     }
 
-    public Stock updateStock(Stock stock) {
-        StockEntity stockEntity = new DozerBeanMapper().map(stock, StockEntity.class);
-        StockEntity entityUpdated = stocksRepository.save(stockEntity);
+    public Stock updateStock(Stock stock) throws Exception{
+        StockEntity stockEntity =  stocksRepository.getById(stock.getStockId());
+        if(Objects.nonNull(stockEntity))
+            throw new RecordNotFoundException("Record not found");
+        StockEntity stockEntityToUpdate = new DozerBeanMapper().map(stock, StockEntity.class);
+        StockEntity entityUpdated = stocksRepository.save(stockEntityToUpdate);
         Stock stockUpdated = new DozerBeanMapper().map(entityUpdated, Stock.class);
         return stockUpdated;
 
     }
 
-    public void deleteStockByStockId(Integer stockId) {
+    public void deleteStockByStockId(Integer stockId) throws  Exception {
+        StockEntity stockEntity =  stocksRepository.getById(stockId);
+        if(Objects.nonNull(stockEntity))
+            throw new RecordNotFoundException("Record not found");
         stocksRepository.deleteById(stockId);
     }
